+++ src/backend/resolvers/mutations/user/update.js
@@ -1,7 +1,7 @@
 const _ = require('lodash')
 
-const upsertUsersPermissions = require('./../../../generate-users-permissions/upsertUsersPermissions')
-const upsertUsersSitemaps = require('./../sitemap/upsertUsersSitemaps')
+const upsertUsersPermissions = require('../sitemap/permissions-upsertion/upsertUsersPermissions')
+const upsertUsersSitemaps = require('../sitemap/sitemaps-upsertion/upsertUsersSitemaps')
 
 const updateUser = async (
   parent,